Move deprecated items to dedicated module
